+++ graphics/CHANGELOG.md
@@ -7,10 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Light::set_pos() and set_size().
+- MapZLayer to set layer Z, Default uses the old static set layers.
 
 ### Changed
 - Light: replaced z with Pos and added Size.
 - Rect::new, Expand to add Color.
+- Map data now remove from tile and set to uniform. Should boost map position update speeds.
 
 ### Removed 
 - Light::set_z().